Optimizations _ v1
Me%lFlowBoundary not being used, so for now was commented
Me%lFlowDischarg e Me%iFlowDischarge alocated only if there are discharges
Me%myWaterColumnOld updated directly in SetBasinColumnToRunoff, and commented the setmatrix value in Modify runoff where it did the same thing. Separated it from when Calibrating1D option was true.
Me%iFlowRouteDFour allocated and used only if Me%RouteDFourPoints was true
Changed the order of the construct so readfile comes before allocate variables