Optimizations V39
Merged part of check stability with the UpdateWaterLevels routine. Added active points matrix to improve  cpu usage. Active points are all points with a flow entering or exiting (updated in dynamicwave only) and then in updatewaterlevels (it is 1 if there is water in a cell).
Used these points to reduce computational cost in integrateFlow.
Volumes are now updated directly in dynamicwave functions + discharges etc instead of in the UpdateWaterLevels